refactor: update code based on comments
