+++ plugins/nemo-experimentalist/tests/experimentalist/test_evaluator_harbor.py
@@ -11,25 +11,27 @@ from types import SimpleNamespace
 
 import pytest
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor import (
-    _TRACE_ARTIFACT_DESTINATION,
-    _TRACE_ARTIFACT_SOURCE,
+    DEFAULT_TRACE_ARTIFACT_SOURCE,
     HarborDataset,
     HarborDependencyContext,
     HarborDependencyRuntime,
-    HarborEvaluator,
-    HarborEvaluatorConfig,
     HarborVerifierValidationError,
     _chmod_path_chain,
-    _cleanup_scoped_imports,
-    _ensure_package,
     _python_syntax_failure,
-    _safe_identifier,
-    _scoped_import_path,
     _shell_syntax_failure,
     _trial_error,
     _trial_metric_spec,
     _trial_metrics,
     _trial_resources,
+)
+from nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native import (
+    _TRACE_ARTIFACT_DESTINATION,
+    HarborEvaluator,
+    HarborEvaluatorConfig,
+    _cleanup_scoped_imports,
+    _ensure_package,
+    _safe_identifier,
+    _scoped_import_path,
     _with_trace_artifact,
 )
 from nemo_experimentalist_plugin.experimentalist.components.evaluator.models import (
@@ -720,7 +722,7 @@ async def test_harbor_evaluator_runs_job_and_maps_output(tmp_path: Path, monkeyp
         async def run(self):
             return SimpleNamespace(id="job-id", stats=FakeStats())
 
-    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor.Job", FakeJob)
+    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native.Job", FakeJob)
 
     result = await evaluator.run(
         agent=tmp_path / "agent",
@@ -845,7 +847,7 @@ async def test_harbor_evaluator_rejects_invalid_python_verifiers_before_job_crea
 
     dataset = HarborDataset.from_path(dataset_dir)
     fake_job = _recording_job(tmp_path / "jobs" / "preflight")
-    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor.Job", fake_job)
+    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native.Job", fake_job)
     compile_calls = 0
     original_compile = compile
 
@@ -966,7 +968,7 @@ async def test_harbor_evaluator_accepts_valid_python_verifier(
     _write(task_dir / "tests" / "check.py", "def check():\n    return True\n")
     dataset = HarborDataset.from_path(task_dir)
     fake_job = _recording_job(tmp_path / "jobs" / "valid-python")
-    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor.Job", fake_job)
+    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native.Job", fake_job)
 
     trials = await HarborEvaluator()._run(agent_dir, dataset, HarborEvaluatorConfig())
 
@@ -988,7 +990,7 @@ async def test_harbor_evaluator_rejects_invalid_configured_test_sh_before_job_cr
     _write(task_dir / "test" / "test.sh", "if true; then\n  echo broken\n")
     dataset = HarborDataset.from_path(task_dir)
     fake_job = _recording_job(tmp_path / "jobs" / "invalid-shell")
-    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor.Job", fake_job)
+    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native.Job", fake_job)
 
     with pytest.raises(HarborVerifierValidationError) as exc_info:
         await HarborEvaluator()._run(agent_dir, dataset, HarborEvaluatorConfig())
@@ -1013,7 +1015,7 @@ async def test_harbor_evaluator_accepts_valid_legacy_test_sh(
     _write(task_dir / "test" / "test.sh", "if true; then\n  echo valid\nfi\n")
     dataset = HarborDataset.from_path(task_dir)
     fake_job = _recording_job(tmp_path / "jobs" / "valid-shell")
-    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor.Job", fake_job)
+    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native.Job", fake_job)
 
     trials = await HarborEvaluator()._run(agent_dir, dataset, HarborEvaluatorConfig())
 
@@ -1327,8 +1329,8 @@ def test_trial_metric_spec_task_dir_verifier(tmp_path):
 def test_with_trace_artifact_already_has_source():
     from harbor.models.job.config import ArtifactConfig
 
-    existing = ArtifactConfig(source=_TRACE_ARTIFACT_SOURCE, destination="traces")
-    result = _with_trace_artifact([existing], _TRACE_ARTIFACT_SOURCE)
+    existing = ArtifactConfig(source=DEFAULT_TRACE_ARTIFACT_SOURCE, destination="traces")
+    result = _with_trace_artifact([existing], DEFAULT_TRACE_ARTIFACT_SOURCE)
     assert result == [existing]
 
 
@@ -1336,7 +1338,7 @@ def test_with_trace_artifact_already_has_destination():
     from harbor.models.job.config import ArtifactConfig
 
     existing = ArtifactConfig(source="/other", destination=_TRACE_ARTIFACT_DESTINATION)
-    result = _with_trace_artifact([existing], _TRACE_ARTIFACT_SOURCE)
+    result = _with_trace_artifact([existing], DEFAULT_TRACE_ARTIFACT_SOURCE)
     assert result == [existing]
 
 
@@ -1344,9 +1346,9 @@ def test_with_trace_artifact_adds_when_missing():
     from harbor.models.job.config import ArtifactConfig
 
     other = ArtifactConfig(source="/other", destination="other")
-    result = _with_trace_artifact([other], _TRACE_ARTIFACT_SOURCE)
+    result = _with_trace_artifact([other], DEFAULT_TRACE_ARTIFACT_SOURCE)
     assert len(result) == 2
-    assert result[0].source == _TRACE_ARTIFACT_SOURCE
+    assert result[0].source == DEFAULT_TRACE_ARTIFACT_SOURCE
 
 
 def test_trial_error_non_dict():
@@ -1419,7 +1421,7 @@ async def test_harbor_evaluator_force_rerun(tmp_path, monkeypatch):
         async def run(self):
             return SimpleNamespace(id="job-id", stats=None)
 
-    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor.Job", FakeJob)
+    monkeypatch.setattr("nemo_experimentalist_plugin.experimentalist.components.evaluator.harbor_native.Job", FakeJob)
 
     trials = await evaluator._run(
         agent=agent_dir,
@@ -1880,10 +1882,10 @@ def test_resolve_trial_task_id_fallback_to_trial_base():
 
 
 def test_with_trace_artifact_string_match():
-    result = _with_trace_artifact([_TRACE_ARTIFACT_SOURCE], _TRACE_ARTIFACT_SOURCE)
-    assert result == [_TRACE_ARTIFACT_SOURCE]
+    result = _with_trace_artifact([DEFAULT_TRACE_ARTIFACT_SOURCE], DEFAULT_TRACE_ARTIFACT_SOURCE)
+    assert result == [DEFAULT_TRACE_ARTIFACT_SOURCE]
 
 
 def test_with_trace_artifact_string_destination_match():
-    result = _with_trace_artifact([_TRACE_ARTIFACT_DESTINATION], _TRACE_ARTIFACT_SOURCE)
+    result = _with_trace_artifact([_TRACE_ARTIFACT_DESTINATION], DEFAULT_TRACE_ARTIFACT_SOURCE)
     assert result == [_TRACE_ARTIFACT_DESTINATION]
